[udev][sata] support scatter/gather
Support discontiguous iotxns and increase the PRD table size to 8k,
so the maximum worst-case transfer size is 32M for a fully
distiguous iotxn. Add a field in block_info_t to return this info.

+++ system/udev/virtio/block.h
@@ -24,7 +24,7 @@ public:
     virtual void IrqConfigChange();
 
     uint64_t GetSize() const { return config_.capacity * config_.blk_size; }
-    uint64_t GetBlockSize() const { return config_.blk_size; }
+    uint32_t GetBlockSize() const { return config_.blk_size; }
     uint64_t GetBlockCount() const { return config_.capacity; }
 
 private:
